test/espnex/layers/test_stft.py
from functools import partial

import jax

from espnex.layers.stft import Stft
from jax.random import PRNGKey, split, uniform, randint
from espnet2.layers.stft import Stft as TStft
import torch
import jax.numpy as jnp
from numpy.testing import assert_allclose


def test_stft():
    key = PRNGKey(0)
    in_shape = bs, samples, channel = 5, 2048, 2

    in_shape = in_shape[:-1]

    key, = split(key, 1)
    ilens = randint(key, [bs], 1000, samples)
    x = uniform(key, in_shape, minval=0, maxval=10000)
    print(x.dtype)

    stft = Stft()
    apply = jax.jit(lambda x, ilens: stft.apply({}, x, ilens))
    y, olens = apply(x, ilens)

    tx = torch.Tensor(jax.device_get(x))
    tilens = torch.Tensor(jax.device_get(ilens))
    tstft = TStft()
    ty, tolens = tstft(tx, tilens)

    assert y.shape == tuple(ty.shape)[:-1]
    assert olens.tolist() == tolens.tolist()

    y = jnp.absolute(y)
    ty = torch.linalg.norm(ty, dim=-1)

    # assert_allclose(jax.device_get(y), ty.numpy(), rtol=.01, atol=.1)
    # diff is caused by the default `reflect` padding of torch.stft, jax do not have this options and I do not see the
    # reason using reflect padding
    pass


